moved app logic to project root
