+++ main.go
@@ -1,7 +1,10 @@
 package main
 
+//https://golang.org/doc/articles/wiki/
 import (
 	"fmt"
+	"log"
+	"net/http"
 	"time"
 )
 
@@ -9,6 +12,18 @@ func main() {
 	if cowntdown(time.Now()) == 0 {
 		fmt.Println("Merry, Christmas!")
 	}
+	http.HandleFunc("/", handler)
+	log.Fatal(http.ListenAndServe(":8080", nil))
+}
+
+/*
+http://localhost:8080/sachiko
+でアクセスすると
+Hi, Merry Christmas sachiko!🎅🎄✨
+が表示される
+*/
+func handler(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprintf(w, "Hi, Merry Christmas %s!🎅🎄✨", r.URL.Path[1:])
 }
 
 func cowntdown(t time.Time) int {
